Initial structure of the Podcast sync system
This doesn't do anything yet, but it implements a good way to communicate
between a service lauched on demand and the UI.

We use an IntentService and a BroadcastReceiver to communicate between Service
and Activity while correctly handling the Activity livecycle.

The original AsyncTask design required the UI to stop the task, rebind it to
a new Activity and restart it every time the receiver Activity is restarted.