Prevent GPodderSync from deadlocking
GPodderSync uses a semaphore to block until the service is bound.

However the callback from the service informing the appliction about the
binding is called in the main thread. If a service call is issued from the
main thread we have a deadlock because we block the thread and the service's
callback can't execute.

Each service call now runs in a separate thread to keep the main one unblocked
and open for the callback.