Fix Bug #17138 Filter does not work on rotation
Make EpisodeFilters serializable and store them if the activity is
recreated. Also update the UI elements that display the
filter status (e.g. the check box in the menu item if the
NewFilter is activiated). I also tried around to prefill
the search term box if a filter is active, because right
now the current search term is reset if you click on the
search icon. Even after a while of trying i could not find a
workardound for this - it seems like this is the way it should
be, i.e. if you click on search the old search data is wiped
and a new search starts.