Fixed bug in streaming player - streaming stop
Previous state: When the player closed detlef and had a streaming episode,
the streaming wouldn't stop. Current state: If the player is currently
playing and streaming, the streaming doesn't stop. If the player is paused
and streaming, the streaming stops. If the user starts it using the notif.
bar again, the streaming starts. If the user pauses it using the notif.
bar, the streaming doesn't stop. If the user closes the notif. bar, the
streaming stops.