Alter detlef classpath settings
Without these modifications, I was getting a variety of errors when
loading Detlef on an Android device. Sometimes, the problem was not
being able to load the superclass on MainActivity, sometimes
mygpoclient-java could not be found. To be honest, I have no idea why
exporting these additional libs is suddenly necessary.