Refs #54 - 2nd attempt to fix the sync
Thank you for the suggestion about FeedParser.onStartAtom. I tried to
find out why it didn't assign anything to the feed url. It seems that I can
fix it by commenting out a check, but I have no clue about what this
check is supposed to do.

So I will just create this pull request, instead of pushing it directly to the
working branch.